[Fix] View, ViewModel 파일 분리
